move free var scanner to separate file
